PP-2472 Add autoUpdateEnabled config flag for viziers in cloud backend
Summary:
this is an ask from customer, which will allow them to deploy pixie widely throughout the clusters, until the hot-update feature is in.
this diff just adds the flag in the db, and allows updating/getting the flag through the UpdateVizierConfig/GetVizierInfo.
It currently does not actually use the flag yet, which will come in a followup

Test Plan: unit tests

Reviewers: zasgar, vihang, nserrino, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D7502

GitOrigin-RevId: 012c6ce9363af0b0f30e871a00e52836155781bd