PC-992 Add usersettings table with analyticsOptout
Summary:
we are reformatting the usersettings table to be more like the userattributes table.
the logic for usersettings and userattributes are the same (minus substituting tourSeen with analyticsOptout), so all of the necessary changes were made in this diff since its just copy & paste.
let me know if it will be easier to just split it up anyway.

Test Plan: launched dev cloud

Reviewers: vihang

Reviewed By: vihang

JIRA Issues: PC-992

Differential Revision: https://phab.corp.pixielabs.ai/D9127

GitOrigin-RevId: 7aa2f0de2c11f5f427df4586faba166e494b97e9