PC-1180: Add requestor IP and responder IP to service graph widget.
Summary:
We want requestor IP and responder IP to show up on the service graph widget for external traffic (a.k.a traffic into or out of the K8s cluster).
This diff adds those fields to the vis spec, and updates the PxL scripts using the service graph widget to produce those values. Note that everything will still work even if those values are not produced, so it's not breaking either way from the perspective of the UI.

Test Plan: ran the scripts, ran the CLI to make sure that proto stays happy

Reviewers: michelle, oazizi, vihang, zasgar

Reviewed By: oazizi

JIRA Issues: PC-1180

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D9864

GitOrigin-RevId: bdae78cc266a078e73db2d9be205fc3ce5cc823b