PC-1305, PC-1310: Support Terminated pod phase in Pixie Cloud and UI
Summary: We currently use the same phases for pods as Kubernetes. However, Kubernetes does not track terminated pods like we do, so terminated pods register as Running in Vizier metadata and Autocomplete. This is confusing to the end user of this data, since the pod is no longer running. In this diff, we add a phase to Pixie Cloud and the Pixie UI that handles a new pod phase, Terminated. This is for pods that neither succeeded nor failed, they were simply terminated by the system. Currently, no Vizier will actually emit any pod updates with this phase, that will come in a future diff. After both of these changes, both autocomplete and queries on pod phase will correctly return the Terminated state for terminated pods, rather than erroneously returning "Running".

Test Plan: will do an end to end test with Vizier changes.

Reviewers: michelle, philkuz

Reviewed By: michelle

JIRA Issues: PC-1305, PC-1310

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10318

GitOrigin-RevId: a85d90cce66fddaefad5b8b7bee47c12caba8435