PC-1436 Add cloud GRPC API for Plugin service
Summary:
Previous diff added the plugin API to gql, so that we can populate the plugins UI in the admin page.
All of our gql endpoints are backed by a GRPC API in cloudpb. This diff adds the GRPC API which matches D10986

Test Plan: We don't use any of these protos in this diff yet, so there is no functionality to test.

Reviewers: vihang, nserrino

Reviewed By: vihang

JIRA Issues: PC-1436

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10994

GitOrigin-RevId: 17128ca1f9de508f96f4a18439fd69682510e112