PC-1436 Fix fields in plugin API
Summary: As I started implementing the plugin grpc service, I realized that there was something that wasn't quite right with my original API. As per the design, users may choose to enable only certain functions in the plugin (for example, only data retention but not alerting). Updated the plugin so that it can be enabled for only certain functions.

Test Plan: these protos are not used anywhere yet and do not affect any existing functionality.

Reviewers: vihang

Reviewed By: vihang

JIRA Issues: PC-1436

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11005

GitOrigin-RevId: 0f624da608995eab6ef01c44bff65ba95523440c