Add customDeployKeySecret to vizier CRD
Summary:
When users are deploying via helm charts, they often don't want to hardcode the deployKey in the helm command/values.yaml.
In most cases, they want to utilize a key that is already specified as a K8s secret, either manually created or created from a secret manager. This secret name is usually something custom, rather than the expected "pl-deploy-secrets" we always use.
This diff adds "customDeployKeySecret" to the CRD so that users can specify the secret name for the deploy key themselves.

Test Plan: This diff doesn't use these fields yet, so there is nothing to test.

Reviewers: vihang, nserrino

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11012

GitOrigin-RevId: ea465d7ce0d5927e36275f34abfd34a58b17a4dd