Update plugin protos to allow for insecure TLS mode
Summary:
This diff updates the plugin protos to include whether insecure TLS should be allowed, and whether the user wants to toggle it or not.
This usecase will be common for users who stand up an OTel collector (or other OSS tool) in their cluster but don't want to deal with SSL certs/productionizing their workload yet.

Test Plan: These are new proto fields that are unused, so no need for testing yet.

Reviewers: vihang, philkuz, nserrino

Reviewed By: philkuz

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11220

GitOrigin-RevId: 6ccffefee7c33ece6ace3a89680f226d9c2ae5af