[GH-582] Adding RPC for generating OTel scripts from px.display scripts
Summary:
We want to generate OTel scripts automatically from scripts that output DataFrames.
This diff adds the interface for creating these scripts.

Users will send over the script and the cluster_id that should be used to generate the script.
Then if the generator has no issues with the script, the user will receive a new script back.

Test Plan: n/a, interface addition with no implementation.

Reviewers: michelle, vihang, zasgar, nserrino

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12128

GitOrigin-RevId: 64d99dc4b7a288dd4eea3302c61fa838def54bc2