[Golang API] Add support for resuming execution after transient failures of the GRPC connection.
Summary:
The Query Broker is setup to accept ExecuteScript requests that are pointed at a particular query ID.
This diff adds support to the Golang API for resuming queries after the grpc connection fails with Unavailable or RST_STREAM.
The retry/resumption respects the original context the user passed into ExecuteScript, and as long as that context is valid, it will retry indefinitely.

Test Plan: Tested that query resumption works by deploying a vizier with Read/WriteTimeouts set to 60s on the grpc servers, and then seeing that the query failed and then was resumed after 60s.

Reviewers: michelle, vihang, philkuz, zasgar

Reviewed By: philkuz

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12305

GitOrigin-RevId: 1588c30d50651e00d25eba976e0ffa66f2697b24