Handle K8sVersion in API service (#862)
Summary: Previous changes added the K8sVersion to the protos... but I
realized the capitalization was off.
This updates the capitalization for the go proto, and also updates the
API service so that it passes the K8sVersion to the configmanager
service.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Updated unit test

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>
GitOrigin-RevId: 92ed07edb77ff3a250b5450f9df77f83db40c04d