Allow timeout to be specified in OTel endpoint config (#897)
Summary: #889 adds a timeout to the OTel collector request. This PR
allows the user to specify the timeout, which is propagated to the plan
node. Once this PR is merged, #889 can be updated to use the timeout
from the plan node, and if none is specified, uses a default.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Updated unit tests

---------

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>
GitOrigin-RevId: 1eb505e20caa5d13f74ccf7a9e15cd8e08cbca23