Upgrade deps for pxapi (#1191)
Summary: This aligns the depedencies exported as part of the pxapi go
module to line up with the dependencies used by the main pixie repo.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Checked out a version of the pxapi package, upgraded the deps
and tested a couple of simple example scripts that use pxapi.

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>
GitOrigin-RevId: f6d98b8327a1df4917d16e46127513fffcf3e8e2