[artifacts] Add artifact mirrors to Artifact proto definition. (#1499)
Summary: Adds `AvailableArtifactMirrors` to the `Artifact` proto
definition. This field will be used as a map from `ArtifactType` to a
list of URLs where that artifact type can be downloaded from. This is
the first step along the way to no longer hardcoding the artifact
download link in the artifact tracker server. This PR also updates the
`artifacts/manifest` package to handle the new field in the proto.

Type of change: /kind cleanup

Test Plan: Added tests in `artifacts/manifest` that the
`ArtifactManifest` type handles the new proto correctly. Other than
that, the new proto is still backwards compatible, so no other code
changes.

---------

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>
GitOrigin-RevId: 099b36e00249bbf39a58c8b54fa7fcaa3f3515b8