Revert " Allow VizierID to be used to find the Vizier's org" (#1666)
Summary: After testing the interaction between the operator and config
manager service, it seems that the config manager service segmentation
faults when it receives the VizierID from the operator as it's not able
to properly convert it to type uuid, causing the service to restart.
This needs to be fixed and in the meantime these changes should be
reverted.

Relevant Issues: #1632

Type of change: /kind bug

Test Plan: Make sure it still builds

Signed-off-by: Kartik Pattaswamy <kpattaswamy@pixielabs.ai>
GitOrigin-RevId: 69415c0bd8af3b3d3aa01e45edc9b76c5261e6f2