Change VizierID type from string to uuidpb.UUID (#1677)
Summary: This PR adjusts the cloud side to accept a type uuidpb.UUID for
the VizierID instead of a string.

Relevant Issues: #1632

Type of change: /kind bug

Test Plan: Skaffold the cloud changes on the testing cluster and then
skaffold the operator changes on a dev cluster. Manually trigger an
update on the operator side to send the VizierID to the config manager
service. Check the config manager service's logs to see that the new
logic is being executed as expected to handle the VizierID.

Signed-off-by: Kartik Pattaswamy <kpattaswamy@pixielabs.ai>
GitOrigin-RevId: e5d83b73a855b85f485dd47e30db437f74fc7dd2